backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "itsdangerous>=2.2.0",
    "notion-client>=2.3.0",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.5",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "qdrant-client==1.8.1",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "status>=0.2.5",
    "uvicorn[standard]>=0.34.3",
    "alembic>=1.16.1",
    "pytest>=8.4.0",
    "cohere>=5.15.0",
    "groq>=0.28.0",
    "cryptography>=45.0.4",
    "inflect>=7.5.0",
]
